+++ buildSrc/src/main/kotlin/Config.kt
@@ -62,7 +62,7 @@ val HOST_OS: OperatingSystem = findHostOs()
 
 object Realm {
     val ciBuild = (System.getenv("CI") != null)
-    const val version = "3.0.0-SNAPSHOT"
+    const val version = "3.1.0-SNAPSHOT"
     const val group = "io.realm.kotlin"
     const val projectUrl = "https://realm.io"
     const val pluginPortalId = "io.realm.kotlin"
@@ -97,47 +97,48 @@ object Realm {
 
 object Versions {
     object Android {
-        const val minSdk = 16
+        const val minSdk = 21
         const val targetSdk = 33
-        const val compileSdkVersion = 33
-        const val buildToolsVersion = "33.0.0"
-        const val buildTools = "7.3.1" // https://maven.google.com/web/index.html?q=gradle#com.android.tools.build:gradle
-        const val ndkVersion = "23.2.8568313"
-        const val r8 = "8.3.37" // See https://developer.android.com/build/kotlin-support
+        const val compileSdkVersion = 35
+        const val buildToolsVersion = "35.0.0"
+        const val buildTools = "8.7.3" // https://maven.google.com/web/index.html?q=gradle#com.android.tools.build:gradle
+        const val ndkVersion = "27.0.12077973"
+        const val r8 = "8.7.18" // See https://developer.android.com/build/kotlin-support
     }
-    const val androidxBenchmarkPlugin = "1.2.0-alpha12" // https://maven.google.com/web/index.html#androidx.benchmark:androidx.benchmark.gradle.plugin
-    const val androidxStartup = "1.1.1" // https://maven.google.com/web/index.html?q=startup#androidx.startup:startup-runtime
-    const val androidxJunit = "1.1.3" // https://maven.google.com/web/index.html#androidx.test.ext:junit
-    const val androidxTest = "1.4.0" // https://maven.google.com/web/index.html#androidx.test:rules
+    const val androidxBenchmarkPlugin = "1.3.3" // https://maven.google.com/web/index.html#androidx.benchmark:androidx.benchmark.gradle.plugin
+    const val androidxStartup = "1.2.0" // https://maven.google.com/web/index.html?q=startup#androidx.startup:startup-runtime
+    const val androidxJunit = "1.2.1" // https://maven.google.com/web/index.html#androidx.test.ext:junit
+    const val androidxTest = "1.6.1" // https://maven.google.com/web/index.html#androidx.test:rules
+
     // Must be built with same (major.minor!?) kotlin version as 'kotlin' variable below, to be binary compatible with kotlin
     const val atomicfu = "0.18.5" // https://github.com/Kotlin/kotlinx.atomicfu
     const val autoService = "1.0" // https://mvnrepository.com/artifact/com.google.auto.service/auto-service
     const val buildkonfig = "0.13.3" // https://github.com/yshrsmz/BuildKonfig
-    const val coroutines = "1.7.0" // https://mvnrepository.com/artifact/org.jetbrains.kotlinx/kotlinx-coroutines-core
-    const val datetime = "0.4.0" // https://github.com/Kotlin/kotlinx-datetime
+    const val coroutines = "1.10.1" // https://mvnrepository.com/artifact/org.jetbrains.kotlinx/kotlinx-coroutines-core
+    const val datetime = "0.6.1" // https://github.com/Kotlin/kotlinx-datetime
     const val detektPlugin = "1.23.6" // https://github.com/detekt/detekt
-    const val dokka = "1.9.0" // https://github.com/Kotlin/dokka
+    const val dokka = "2.0.0" // https://github.com/Kotlin/dokka
     const val gradlePluginPublishPlugin = "0.15.0" // https://plugins.gradle.org/plugin/com.gradle.plugin-publish
     const val jmh = "1.34" // https://github.com/openjdk/jmh
     const val jmhPlugin = "0.6.6" // https://github.com/melix/jmh-gradle-plugin
     const val junit = "4.13.2" // https://mvnrepository.com/artifact/junit/junit
     const val kbson = "0.4.0" // https://github.com/mongodb/kbson
     // When updating the Kotlin version, also remember to update /examples/min-android-sample/build.gradle.kts
-    const val kotlin = "2.0.20" // https://github.com/JetBrains/kotlin and https://kotlinlang.org/docs/releases.html#release-details
-    const val kotlinJvmTarget = "1.8" // Which JVM bytecode version is kotlin compiled to.
-    const val latestKotlin = "2.0.20" // https://kotlinlang.org/docs/eap.html#build-details
-    const val kotlinCompileTesting = "0.5.1" // https://github.com/zacsweers/kotlin-compile-testing
-    const val ktlint = "0.45.2" // https://github.com/pinterest/ktlint
+    const val kotlin = "2.1.0" // https://github.com/JetBrains/kotlin and https://kotlinlang.org/docs/releases.html#release-details
+    const val kotlinJvmTarget = "17" // Which JVM bytecode version is kotlin compiled to.
+    const val latestKotlin = "2.1.0" // https://kotlinlang.org/docs/eap.html#build-details
+    const val kotlinCompileTesting = "0.7.0" // https://github.com/zacsweers/kotlin-compile-testing
+    const val ktlint = "1.5.0" // https://github.com/pinterest/ktlint
     const val ktor = "2.3.12" // https://github.com/ktorio/ktor
     const val multidex = "2.0.1" // https://developer.android.com/jetpack/androidx/releases/multidex
     const val nexusPublishPlugin = "1.3.0" // https://github.com/gradle-nexus/publish-plugin
-    const val okio = "3.9.0" // https://square.github.io/okio/#releases
+    const val okio = "3.10.2" // https://square.github.io/okio/#releases
     const val relinker = "1.4.5" // https://github.com/KeepSafe/ReLinker
-    const val serialization = "1.7.1" // https://kotlinlang.org/docs/releases.html#release-details
-    const val shadowJar =  "6.1.0" // https://mvnrepository.com/artifact/com.github.johnrengelman.shadow/com.github.johnrengelman.shadow.gradle.plugin?repo=gradle-plugins
+    const val serialization = "1.8.0" // https://kotlinlang.org/docs/releases.html#release-details
+    const val shadowJar =  "8.1.0" // https://mvnrepository.com/artifact/com.github.johnrengelman.shadow/com.github.johnrengelman.shadow.gradle.plugin?repo=gradle-plugins
     const val snakeYaml = "1.33" // https://github.com/snakeyaml/snakeyaml
-    val sourceCompatibilityVersion = JavaVersion.VERSION_1_8 // Language level of any Java source code.
-    val targetCompatibilityVersion = JavaVersion.VERSION_1_8 // Version of generated JVM bytecode from Java files.
+    val sourceCompatibilityVersion = JavaVersion.VERSION_17 // Language level of any Java source code.
+    val targetCompatibilityVersion = JavaVersion.VERSION_17 // Version of generated JVM bytecode from Java files.
 }
 
 // Could be actual Dependency objects