+++ packages/plugin-compiler/src/main/kotlin/io/realm/kotlin/compiler/RealmModelSyntheticPropertiesGeneration.kt
@@ -92,6 +92,7 @@ import org.jetbrains.kotlin.ir.expressions.impl.IrConstructorCallImpl
 import org.jetbrains.kotlin.ir.expressions.impl.IrGetEnumValueImpl
 import org.jetbrains.kotlin.ir.expressions.impl.IrPropertyReferenceImpl
 import org.jetbrains.kotlin.ir.expressions.impl.IrVarargImpl
+import org.jetbrains.kotlin.ir.expressions.impl.fromSymbolOwner
 import org.jetbrains.kotlin.ir.symbols.UnsafeDuringIrConstructionAPI
 import org.jetbrains.kotlin.ir.types.IrSimpleType
 import org.jetbrains.kotlin.ir.types.IrType
@@ -303,7 +304,7 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                 type = companionFieldsType,
                 symbol = mapOf,
                 typeArgumentsCount = 2,
-                valueArgumentsCount = 1,
+//                valueArgumentsCount = 1,
                 origin = null,
                 superQualifierSymbol = null
             ).apply {
@@ -333,15 +334,14 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                 symbol = propertyElementType.classOrNull!!,
                                 classType = propertyElementType.classOrNull!!.defaultType,
                             )
-                            val objectPropertyType = if (it.value.isComputed) realmObjectPropertyType else
-                                realmObjectMutablePropertyType
+                            val objectPropertyType = if (it.value.isComputed) realmObjectPropertyType else realmObjectMutablePropertyType
                             val elementType = pairClass.typeWith(pluginContext.irBuiltIns.kClassClass.typeWith(), objectPropertyType)
                             // Pair<String, Pair<String, KMutableProperty1<*, *>>>()
                             IrConstructorCallImpl.fromSymbolOwner(
                                 startOffset = startOffset,
                                 endOffset = endOffset,
-                                type = elementType,
-                                constructorSymbol = pairCtor
+                                type = elementType as IrType,
+                                constructorSymbol = pairCtor as org.jetbrains.kotlin.ir.symbols.IrConstructorSymbol
                             ).apply {
                                 putTypeArgument(0, pluginContext.irBuiltIns.stringType)
                                 putTypeArgument(1, elementType)
@@ -359,8 +359,8 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                     IrConstructorCallImpl.fromSymbolOwner(
                                         startOffset = startOffset,
                                         endOffset = endOffset,
-                                        type = elementType,
-                                        constructorSymbol = pairCtor
+                                        type = elementType as IrType,
+                                        constructorSymbol = pairCtor as org.jetbrains.kotlin.ir.symbols.IrConstructorSymbol
                                     ).apply {
                                         putTypeArgument(
                                             0,
@@ -507,7 +507,7 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                             type = classInfoClass.defaultType,
                             symbol = classInfoCreateMethod.symbol,
                             typeArgumentsCount = 0,
-                            valueArgumentsCount = 5
+//                            valueArgumentsCount = 5
                         ).apply {
                             dispatchReceiver = irGetObject(classInfoClass.companionObject()!!.symbol)
                             var arg = 0
@@ -714,7 +714,7 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                 } ?: irNull(pluginContext.irBuiltIns.kClassClass.typeWith(typedRealmObjectInterface.defaultType).makeNullable())
 
                                 // Define the link target. Empty string if there is none.
-                                val linkPropertyName: IrConst<String> = if (type == linkingObjectType) {
+                                val linkPropertyName: IrConst = if (type == linkingObjectType) {
                                     val targetPropertyName = getLinkingObjectPropertyName(backingField)
                                     irString(targetPropertyName)
                                 } else {
@@ -727,7 +727,7 @@ class RealmModelSyntheticPropertiesGeneration(private val pluginContext: IrPlugi
                                     type = propertyClass.defaultType,
                                     symbol = propertyCreateMethod,
                                     typeArgumentsCount = 0,
-                                    valueArgumentsCount = 10
+//                                    valueArgumentsCount = 10
                                 ).apply {
                                     var arg = 0
                                     // Persisted name