+++ packages/plugin-compiler/src/main/kotlin/io/realm/kotlin/compiler/SyncLoweringExtension.kt
@@ -140,7 +140,7 @@ private class SyncLowering(private val pluginContext: IrPluginContext, private v
                     type = expression.type,
                     symbol = target.symbol,
                     typeArgumentsCount = 0,
-                    valueArgumentsCount = target.valueParameters.size,
+//                    valueArgumentsCount = target.valueParameters.size,
                     origin = null,
                     superQualifierSymbol = null
                 ).apply {